+++ src/main/java/com/sht/eurasiaring/entity/Reply.java
@@ -17,16 +17,16 @@ import java.util.List;
 public class Reply implements Serializable {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 	private String createTime;
 	private String content;
 	private Integer number;
-	private Long leaf;   //等于0时就是一个树的末尾
-	private Long userId;  //发布回复的人id
-	private Long matterId;  //事件的id
-	private Long commentId;
-	private Long nameId;  //给谁回复的人id 父id
-	private Long parentId; // reply的父id
+	private Integer leaf;   //等于0时就是一个树的末尾
+	private Integer userId;  //发布回复的人id
+	private Integer matterId;  //事件的id
+	private Integer commentId;
+	private Integer nameId;  //给谁回复的人id 父id
+	private Integer parentId; // reply的父id
 
 	@Transient
 	private String parentName;//父name
