+++ src/main/java/com/sht/eurasiaring/repository/CarouselRepository.java
@@ -8,5 +8,5 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
  * @author Hongtao Shen
  * @date 2020/5/16 - 18:17
  **/
-public interface CarouselRepository extends JpaRepository<Carousel, Long>, JpaSpecificationExecutor<Carousel> {
+public interface CarouselRepository extends JpaRepository<Carousel, Integer>, JpaSpecificationExecutor<Carousel> {
 }
