+++ src/main/java/com/sht/eurasiaring/repository/ClassifyRepository.java
@@ -8,5 +8,5 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
  * @author Hongtao Shen
  * @date 2020/5/16 - 17:33
  **/
-public interface ClassifyRepository extends JpaRepository<Classify, Long>, JpaSpecificationExecutor<Classify> {
+public interface ClassifyRepository extends JpaRepository<Classify, Integer>, JpaSpecificationExecutor<Classify> {
 }
