+++ src/main/java/com/sht/eurasiaring/repository/UserRepository.java
@@ -4,7 +4,7 @@ import com.sht.eurasiaring.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface UserRepository extends JpaRepository<User, Long>, JpaSpecificationExecutor<User> {
+public interface UserRepository extends JpaRepository<User, Integer>, JpaSpecificationExecutor<User> {
 
     public User findByOpenid(String openid);
 }
